refactor(teacher-see-chatGPT): move evaluation related model in the same parent folder
