+++ src/main/kotlin/org/elaastic/questions/player/components/evaluation/chatGptEvaluation/ChatGptEvaluationModel.kt
@@ -1,4 +1,4 @@
-package org.elaastic.questions.player.components.chatGptEvaluation
+package org.elaastic.questions.player.components.evaluation.chatGptEvaluation
 
 import org.elaastic.questions.assignment.sequence.ReportReason
 import org.elaastic.questions.assignment.sequence.UtilityGrade
