+++ src/main/kotlin/org/elaastic/questions/player/phase/result/LearnerResultPhaseViewModel.kt
@@ -1,7 +1,6 @@
 package org.elaastic.questions.player.phase.result
 
-import ognl.Evaluation
-import org.elaastic.questions.player.components.chatGptEvaluation.ChatGptEvaluationModel
+import org.elaastic.questions.player.components.evaluation.chatGptEvaluation.ChatGptEvaluationModel
 import org.elaastic.questions.player.components.results.ResultsModel
 import org.elaastic.questions.player.components.studentResults.LearnerResultsModel
 import org.elaastic.questions.player.phase.PhaseViewModel
